// src/app/types.ts
export interface AboutUsDataProps {
  title: string;
  introduction: string;
  introduction1: string;
  video: VideoDataProps;
  cta: CTADataProps;
}

export interface BreadCrumbsDataProps {
  backgroundImage: string;
  title: string;
  breadcrumbs: BreadCrumbsLinkItems[];
}

export interface BreadCrumbsLinkItems {
  label: string;
  link: string;
  active: boolean;
}

export interface BrandDataProps {
  brands: string[];
}
// ...
export interface CTADataProps {
  backgroundImage: string;
  title: string;
  buttonText: string;
  buttonUrl: string;
}

export interface CardDataProps {
  title: string;
  buttonLink: string;
  backgroundImage: string;
  tags: string[];
}
// ...
export interface FeatureDataProps {
  sectionTitle: string;
  buttonText: string;
  buttonUrl: string;
  image: string;
  features: {
    icon: string;
    title: string;
    description: string;
  }[];
}
// ...
export interface GalleryDataProps {
  imgSrc: string;
  title: string;
  year: string;
  height: string;
}
// ...
export interface TeamMembersDataProps {
  title: string;
  subtitle: string;
  teamMembers: {
    img: string;
    name: string;
    role: string;
    description: string;
    link: string;
  }[];
}

// Testimonial Section Data Type
export interface TestimonialDataProps {
  sectionTitle: string;
  testimonials: {
    text: string;
    name: string;
    designation: string;
    img: string;
  }[];
}

// Video block section
export interface VideoDataProps {
  backgroundImage: string;
  videoUrl: string;
}

// Value Section Data Type
export interface ValueDataProps {
  sectionTitle: string;
  values: { text: string }[];
}

// Works Section Data Type
export interface WorksDataProps {
  title: string;
  subtitle: string;
  slides: {
    image: string;
    name: string;
    location: string;
    description: string;
  }[];
}
// Works Section Two Data Type
export interface WorksTwoDataProps {
  title: string;
  subtitle: string;
  galleryItems: GalleryDataProps[];
}